Setup Intern.js client for unit tests and coverage reporting
