+++ package.json
@@ -8,6 +8,7 @@
     "grunt-contrib-connect": "^0.11.2",
     "grunt-contrib-jshint": "^0.10.0",
     "grunt-contrib-uglify": "^0.5.0",
+    "grunt-open": "^0.2.3",
     "intern": "^3.0.6",
     "jshint": "^2.5.2"
   },
